Improves sidebar handling in Firefox
After vertical tab support PR has been merged, this caused issues in my
local Firefox configuration. I use the Tree Style Tab extension, and
have the sidebar open on the right side. So with those changes the
Chatterino window goes off to the right from the window, as the code
assumes there will be sidebars only on the left.

There doesn't seem to be a good cross-compatible way of detecting the
tab viewport offset from the window. See
https://github.com/w3c/csswg-drafts/issues/809. But, luckily, on Firefox
there is still "mozInnerScreenX", which can be used to figure that out.
And since the original fix was only for Firefox, then might as well use
it.

Since this is a DOM property, the xOffset calculation has been moved to
the injected script.

The hacky offset logic was also changed, and there are still questions
regarding what is going on there with the margins, but, in practice, it
seems to be working good enough.

As for zoom logic changes, it looks like on Firefox devicePixelRatio is
not just zoom * displayScaleFactor and there is some hidden multiplier,
which affects the actual zoom value. Added a workaround for this, which
assumes, that display scale factor is a multiple of 25%, so that we
could get the _actual_ zoom value of the page. This seems to improve
zoom handling drastically.